requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==3.2.21
django-allauth==0.41.0
django-countries==7.2.1
django-crispy-forms==1.14.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2==2.9.7
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
stripe==6.5.0
